+++ src/Moneybird.Net/Misc/WorkflowType.cs
@@ -1,8 +1,13 @@
+using System.Text.Json.Serialization;
+
 namespace Moneybird.Net.Misc
 {
+    [JsonConverter(typeof(JsonStringEnumMemberConverter))]
     public enum WorkflowType
     {
+        [JsonPropertyName("InvoiceWorkflow")]
         InvoiceWorkflow,
+        [JsonPropertyName("EstimateWorkflow")]
         EstimateWorkflow
     }
 }