+++ src/Moneybird.Net/MoneybirdClient.cs
@@ -4,6 +4,7 @@ using Moneybird.Net.Endpoints.Administrations;
 using Moneybird.Net.Endpoints.Contacts;
 using Moneybird.Net.Endpoints.CustomFields;
 using Moneybird.Net.Endpoints.DocumentStyles;
+using Moneybird.Net.Endpoints.Users;
 using Moneybird.Net.Endpoints.Workflows;
 using Moneybird.Net.Http;
 
@@ -18,6 +19,7 @@ namespace Moneybird.Net
         public IContactEndpoint Contact { get; }
         public ICustomFieldEndpoint CustomField { get; }
         public IDocumentStyleEndpoint DocumentStyle { get; }
+        public IUserEndpoint User { get; }
         public IWorkflowEndpoint Workflow { get; }
 
         /// <summary>
@@ -48,6 +50,7 @@ namespace Moneybird.Net
             Contact = new ContactEndpoint(_config, requester);
             CustomField = new CustomFieldEndpoint(_config, requester);
             DocumentStyle = new DocumentStyleEndpoint(_config, requester);
+            User = new UserEndpoint(_config, requester);
             Workflow = new WorkflowEndpoint(_config, requester);
         }
     }